backend/requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
DateTime==4.5
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
dnspython==2.2.1
entrypoints==0.3
filelock==3.3.1
Flask==2.2.2
google-api-core==2.10.1
google-auth==2.11.1
google-cloud-appengine-logging==1.1.4
google-cloud-audit-log==0.2.4
google-cloud-core==2.3.2
google-cloud-logging==3.2.2
google-cloud-tasks==2.10.2
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
proto-plus==1.22.1
protobuf==4.21.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pyzmq==22.3.0
requests==2.28.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
twilio==7.14.0
urllib3==1.26.12
virtualenv==20.9.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.8.1
zope.interface==5.4.0
